refactor :move directory structure test
