+++ __tests__/test/validate.js
@@ -1,5 +1,5 @@
-import { validateInput } from '../parsers/Validate.js'; // default export를 중괄호 없이 가져오기
-import { ERROR_MESSAGE } from '../const/index.js';
+import { validateInput } from '../../src/parsers/Validate.js'; // default export를 중괄호 없이 가져오기
+import { ERROR_MESSAGE } from '../../src/const/index.js';
 
 describe('validate test', () => {
   test('만욱,노미,재걸,동호', () => {
